Struct jtex::JupiterDecoder

pub struct JupiterDecoder { /* private fields */ }

The main struct for decoding Jupiter Texture images.

This is intended to be used alongside the [image] crate, as it implements the [ImageDecoder] trait.

Implementations

impl JupiterDecoder

pub fn decode<R>(reader: R) -> Result<JupiterDecoder, Error>where R: Read,

Decode the contents provided by the reader.

This decoder will read the image contents into memory immediately, because the image format does not lend itself to streaming.

Errors

May return any variant of the crate error type, Error.